fix: kubectl cnpg status PDB list (#4530)
In the kubectl plugin, when getting the list of PodDisruptionBudgets,
the code wasn't restricting the selection to the namespace where
the target Cluster resource has been created.

This patch fixes that, with the added benefit of restricting the set
of privileges needed by the user.

Fixes #4522

Signed-off-by: HaveFun83 <38665716+HaveFun83@users.noreply.github.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit 7420aed9754c32085f526cd205c9ba11af6097d6)