fix: timeout when restarting PostgreSQL and while lifting fencing (#4504)
The instance manager starts PostgreSQL:

1. when it starts up
2. when configuration changes are being applied (after stopping it)
3. when fencing is lifted.

In the second and third examples, the operator is requested by the
embedded cluster reconciler loop, and performed without any timeout.

If PostgreSQL won't start up again because of a wrong configuration or
missing disk space, the reconciler loop will be stuck waiting for a dead
postmaster to be up.

This patch handles this condition by using a combination of the timeout
parameters that are already set in the cluster.

Fixes: #4501

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 09a4d80a1783e1e2144cae3954ba090c5225db15)