fix: avoid immediately the instance manager if it is not needed
If we restarted PostgreSQL, but we don't manage to connect to it, we
should proceed with the shutdown logic of the instance manager instead
of quitting immediately with exit code 1.

This is specially important when PG shut down because of missing disk
space and when PG is up but trying to reach a consistent state.

Closes: #4669

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>