fix: avoid shutting down the instance manager if it is not needed (#4670)
If we restart PostgreSQL but don't manage to connect to it, we should
proceed with the instance manager's shutdown logic instead of quitting
immediately with exit code 1.

This is especially important when PG shuts down because of missing disk
space and when PG is up but trying to reach a consistent state.

Closes: #4063

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 99b1ba47ad5af2400526dd1da79789162e99e400)