feat: implement predicate_query to validate which are the metrics to be collected (#4503)
This patch introduces a new field called "predicate_query" in the user-defined metrics. 

The `predicate_query` is an SQL query that returns at most one row and one `boolean` column to run on the target database. The system evaluates the predicate and, if `true,` executes the `query.`

This allows the users to execute the user-defined query only when certain criteria are met.

Closes #4499 

Signed-off-by: Gabriele <gabriele.rossini@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>