test(e2e): fix backup failures for replica mode (#4650)
Changed `Context("can bootstrap a replica cluster from a backup")`
tests so that both object store and volume snapshot subtests can make
use of the same source Cluster. This will save time and will fix the
issue of Backups being stuck in the "started" state.

Changed `AssertReplicaModeCluster()` to allow passing a name for the
test table that will be created so that it can be used multiple time
against the same source Cluster.

Changed `AssertDetachReplicaModeCluster()` to add a new check to
verify the non-existence of the replica cluster's user. This assertion
now also works standalone and not only in conjunction with
`AssertReplicaModeCluster()`.

The previous fixes, while improving the E2e test codebase, fixes the
backup failures we had in the Replica Mode E2E tests.

Closes #4646

Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>