fix: allow raising replication parameters when restoring a backup (#4564)
Ensure the PostgreSQL replication parameters are set to the higher value
between the ones specified in the cluster specification and the ones
stored in the backup.

This will ensure that the backup will be restored correctly while
allowing the users to raise their value to accommodate changes in the
configuration that have happened after the backup was taken.

Partially closes #2478 #2337

Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit 87f80ce9d3468ecccba07a6f50fa2cdb9f614e72)