chore: remove unused systemUID detection (#4735)
We were acquiring the UID of the kube-system namespace to create a so
called systemUID, which hasn't been used for a long time.

Since that value is not really used, we remove it and its detection mechanism.

Closes #4734

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
(cherry picked from commit 5386c931fcee65f90cce5b43ab02be1eec14f5ff)