feat: prevent failovers when disk space is exhausted (#4404)
PostgreSQL will shut down cleanly when there is not enough disk space to
store WAL files.

The operator did not recognize this condition and, since the primary
failed, was performing a failover to the most advanced replica. This
action will not fix the underlying issue.

Only a manual disk resize, initiated by the user, can ultimately lead to
a fully working PostgreSQL cluster.

This patch makes the instance manager recognize this condition and
report it to the operator. Upon detecting it, the operator will not
trigger a switchover and set a phase describing the situation.

After the PVCs are resized, the cluster will restart working correctly.

Closes: #4521

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>