fix: always use `RegisterPhase` to change the phase (#4849)
This patch aims to solve some rare cases in which the cluster phase and
cluster readiness conditions are out of sync.

This can happen because in some parts of the code, we still directly set
the phase instead of passing through the dedicated function
`RegisterPhase`; this patch aims to solve that.

Closes #4209

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>