fix: prevent concurrent postmaster lifecycle loops (#4920)
With the current implementation, it was possible to have two concurrent
postmaster lifecycle loops: the first one could be waiting for the
instance permissions to be set up, the second one could be starting.

This resulted in a race condition in the flag that enabled and
disabling the readiness probe, resulting in the second postmaster
running to be ready while the readiness probe was artificially forced to
false.

This condition is triggered by fencing a cluster and immediately lifting
the fencing.

Closes: #4916

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit d76ef278453805742c4f4e650902662ea733e51f)