feat: add `.spec.replicaCluster.primary` field (#4388)
This patch introduces changes to the handling of the
`.spec.replica.enabled` field, allowing it to be set to `nil`. Prior to
this patch, only `true` and `false` were permitted, with `false` being
the default. When the `.spec.replica.enabled` field is set to `nil`, the
cluster's classification as a primary or replica cluster will be
determined by two new fields, `.spec.replica.primary` and
`.spec.replica.self`.

The `.spec.replica.self` field specifies the name of the current
cluster, which will be compared with the `.spec.externalClusters[].name`
field. The `.spec.replica.primary` field specifies the name of the
current primary cluster, which will be searched in the same list.
Under the new behavior, when the `.spec.replica.enabled` field is set to
`nil` and the `.spec.replica.primary` and `.spec.replica.self` fields
contain the same value, the cluster is classified as a primary cluster.

Conversely, when the `.spec.replica.enabled` field is `nil` and the
values of `.spec.replica.primary` and `.spec.replica.self` differ, the
cluster will be classified as a replica cluster.

The purpose of this patch is to streamline the configuration process for
a distributed PostgreSQL topology, where the `.spec.externalClusters`
section is synchronized across different data centers.

Closes #3942

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>