feat(cnpg-i): support remote plugins
This patch allows the operator to work with remote CNPG-i plugins, that
are expected to be deployed in a different Pod in the same namespace of
the operator.

The discovery of plugins is based on a set of annotations and labels to
be added to services to be used as a plugin.

The plugin developer or distributor is expected to provide a set of
credentials that will be used by the operator to authenticate to the
plugin via mTLS, and vice-versa.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>