feat(managed-services): add support for `updateStrategy` (#4952)
Introduce the `updateStrategy` field to control how the operator updates
a service definition. By default, the operator uses the `patch`
strategy, applying changes directly to the service. Alternatively, the
`recreate` strategy deletes the existing service and recreates it from
the template.

This is particularly useful to handle changes in cloud-based load
balancers for DBaaS scenarios.

Closes #4922

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>