fix(pooler): enhance rollout control during operator upgrade (#5006)
This update clarifies Pooler pods' restart behavior. Previously, the
operator image name was excluded from the configuration hash
calculation, so an operator upgrade would not have generated a rollout
until anything else in the spec changed, even if that change did not
require a rollout. This behavior was surprising for the user.

To remove surprises, we now include the operator image in the
configuration hash, so every upgrade triggers a rollout.

Additionally, for users who prefer more control over rollouts, we've
introduced the annotation `cnpg.io/reconcilePodSpec: disabled`. This
annotation prevents any updates to the Pooler deployment, except for
changes to the number of replicas.

Closes #4999

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>