fix: cleanup PGDATA before initializing a PVC
This commit addresses an issue in CloudNativePG where Kubernetes jobs are
created to initialize new PVCs during replica addition and primary instance
creation.

In case of initialization failure, the jobs are retried, but since
the PGDATA directory already exists, subsequent initialization attempts fail.

The patch modifies the process to clean up the PGDATA directory before retrying
initialization. If the existing directory is a valid PGDATA, it is moved to a
backup location instead of being deleted to prevent accidental data loss in
case of a pre-existing PGDATA mistakenly left by the user in the PV.

Closes: #3365

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>