wip: use the dbname argument in the connection string when connecting to external clusters
The connectionParameters section of an ExternalCluster lets you specify a dbname
that is added to the connection string but that isn't useful when it is a
monolith server hosting many databases where many need to be acted on.

This commit takes the 'dbname' argument and adds it the externa cluster
connection string so commands work in the same way on as they do on local
databases. If set, the argument overrides the dbname set in the external cluster
config.

closes #4140

Signed-off-by: Seumas Dunlop <seumas.dunlop@coralbay.tv>