chore: improve OLM environment detection (#5031)
Look at the `operators.coreos.com/v1` resources and check that
it they exist. If this is the case, we're in an OLM environment.
Before we were relying on the detection of the SecurityContextConstraint
 to know that we were running on an OLM environment but an OLM
environment may or may not have a SCC inside.

This helps us to improve the plugin report and properly retrieve more
useful information.

Closes #3764

---------

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
(cherry picked from commit 91d70bc0ef1f19cac006bfbfbf039f5316e11450)