fix: cleanup PGDATA before initializing a PVC (#5112)
This commit addresses an issue in CloudNativePG concerning the
initialization of new Persistent Volume Claims (PVCs) during the addition
of replicas and the creation of primary instances. In cases where the
initialization fails, Kubernetes jobs are retried. However, a partial
initialization may leave behind a PGDATA directory, causing subsequent
retries to fail.

To avoid this, the patch introduces a procedure to clean up the PGDATA
directory before retrying the initialization. If the directory is
identified as a valid PostgreSQL data directory, it is moved to a
backup location rather than deleted. This precautionary step helps
prevent accidental data loss in situations where the PGDATA directory
might have been mistakenly pre-existing in the Persistent Volume (PV).

Closes: #3365

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>