docs: considerations on block storage solutions (#5036)
This update adds details to the storage documentation, focusing on the
data-locality feature. It highlights scenarios where data instances might
reside on the same node, posing a risk if the node is destroyed (following the
"cattle, not pets" approach). Such situations could lead to cluster corruption.

The documentation outlines these risks and suggests potential solutions to
mitigate them.

Relates #4355

Signed-off-by: Wei-Yen Tan <weiyen.tan@gmail.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>