ci: use full linter version for testing (#5343)
Before this, we were using only major.minor version to install and run
the golangci-lint. In the latest patch version, we found lot of new
issues and this started to appear in the linter due to the action
installing the latest version available for the minor version.
This avoid that by using the full version of the linter to install and
setting it to the previous one. Also configures renovate to open PRs
on linter updates.

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
(cherry picked from commit 297bc77679e3d041c4d97f75e9edc75a2264fad1)