chore: avoid gosec when calling term.IsTerminal()
The function IsTerminal() requires an integer as input, in this case,
the use is calling a file descriptor to "/dev/stdout", the problem is
that every file descriptor is a uintptr type, requiring a cast, this
cast from uintptr to int triggers the latest version of the gosec linter.

We avoid the warning by just ignoring the linter since this is the know
way to detect the terminal now.

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>