chore: change WAL segment struct from int32 to int64
One of the reasons we have int32 in the WAL segment description is
that's the length in PostgreSQL, but all the operations after reading
the segment are int64 and we cast to int32 in a couple of places,
but these cast aren't necessary.

This avoid the gosec linter complaining about the casting from
int64 to int63.

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>