feat: predefined cnpg-i plugins
This patch adds a new parameter to the operator configuration named
`INCLUDE_PLUGINS`, containing a comma-separated list of plugins to be
always included in the cluster reconciliation.

The installation of the list of plugins is done by the mutating webhook,
and the user can deactivate them at cluster-level by using a new field
named `.spec.plugins.enabled`, defaulting to true.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>