test: remove redundant test for cobra library functionality (#5431)
This test does not verify our code; instead, it tests the functionality
of the cobra library, which is outside our scope. Additionally, the test
is unreliable as it inherits arguments from the test command line,
causing unpredictable behavior. It also outputs the help text directly
into the test results, leading to confusing and misleading test output.
Removing this test will help maintain clarity and focus on our codebase.

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>