test(e2e): ensure AssertSwitchoverWithHistory() waits for cluster readiness (#5437)
Update the `AssertSwitchoverWithHistory()` function to include a wait
until the cluster is fully ready after a switchover. This change
addresses potential issues with test timing and execution by ensuring
that tests do not proceed until the cluster is confirmed to be in a
stable state.

Closes #5436

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
(cherry picked from commit 071c4e672042bbc10010b6b37af2493976581034)