fix: ensure orphan services are reconciled after restoring a cluster (#5379)
This patch ensures that the service manifests restored through a
backup/restore tool are correctly reconciled and not left over as
orphans. This is done by deleting the orphan services so they can be
recreated with the proper ownership by the operator.

Before this patch, this caused an error because the
reconcileRestoredCluster refused to reconcile services that the operator
didn't own.

Closes #5373

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit bbe6b69e2ea35aa8be42ef77f07cf9f50c1c6aaf)