chore(tests): remove ignore-fails tests (#5496)
The ignore-fails test was created a long time ago to ensure that the
`ignore-test` test labels were working as intended. However, we never
used it, which indicates that it is not necessary.

Moreover, the new GitHub output from ginkgo links every error, but the
ignore-fail test causes too many annotations in the GitHub Summary,
which reduces the feature's usefulness.

For this reason, we decided to remove the ignore-fails feature and the
corresponding E2E test.

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit a16b29c8709e05e465575c07fb5c60850f0f274c)