feat: declarative database management (#5325)
Add support for declarative database management by introducing a new
Custom Resource called `Database`.

This new CR is controlled by the instance manager of the primary
instance.

Database deletion is controlled via a finalizer, with a default
retention policy of `retain`. As a result, deleting a Database object in
Kubernetes will not remove the underlying PostgreSQL database by
default. The user can change this behavior by explicitly setting the
retention policy to `delete`.

Closes #5501

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>