tests(e2e): catch operator logs during the tests run (#5457)
Since we included the catch all for the logs of all the pods and show
that it works, now we can do the same for the operator logs using the
same functions and model.

We gather the operator pods by matching the labels and we split them by
different pod name, meaning that if the operator is destroyed and
recreated we will have a different log file for the new pod.

Closes #5453

---------

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
(cherry picked from commit dd37ae7ab95e6c21652b1e4dbe557718bfefdf2a)