chore(test): avoid the MinIO pods deployment at the beginning of tests
The deployment of MinIO and `mc` container was using most of the time
at the beginning of the e2e tests making really slow to trigger a non
related tests. This commit fix that by making MinIO a resource that is
asked by the tests that require it and reusing the already existing one
if it was requested before by another tests. It will not save much time
when running backup tests, but it will save a lot of time just for
other non related tests.

This is a first step to reduce the MinIO need into an encapsulated form
for a future easy removal and/or replacement, from now on it will not
be a strong need for the starting and a couple of functions are MinIO
specific.

Another introduced benefits it's a set of function that query the store
instead of trying to deal with all the ways of working of each Object
Store provider. These functions must be cleaned and improved in the future
for a more generic work, but for now it's enough to work with.

Closes #6572

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>