chore(e2e): prevent VolumeSnapshot E2e to fail because of "tar"
E2e tests are using the CSI-hostpath implementation of VolumeSnapshots,
which are based on `tar`.

When taking a snapshot of a running PostgreSQL instance, `tar` may
download the content of a file that was present at the start of the
process but removed later.

With this patch, the setup script is using the `ignoreFailedRead`
VolumeSnapshotClass to prevent the VolumeSnapshot from failing because
of that.

See: https://github.com/kubernetes-csi/csi-driver-host-path/pull/543

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>