feat: remove pg_database_size from status probe
The instance manager status probe was using `pg_database_size` which
scans the whole PGDATA to compute the current size of every database.

Instead of doing that in the status probe, causing a scan in each
reconciliation loop, we now do that only when requested by the `kubectl
cnpg status` plugin.

Closes: #5686

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>