chore: abort certificate creation upon fatal error detection (#5574)
Previously, the certificate renewal process attempted to renew
or create a certificate upon Secret retrieval, even in cases
where the Secret `Get` request failed due to permanent errors
like insufficient privileges.

This led to incorrect error logging and confusion.

With this patch, the certificate renewal process only continues
when the Secret is not found or renewal is required.

Closes #5575

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>