feat(kubectl-cnpg): kubectl cnpg logs pretty
This patch adds a new command named `kubectl cnpg logs pretty`, that
reads a log stream from the standard input and writes to the standard
output a human-readable version of it.

It allows the user to filter the logs for some pods or some logger, and
to avoid showing the logs behind the specified level.

Closes: #5770

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>