fix: avoid concurrent updates on the designed primary (#5755)
This patch fixes a condition where a demoted replica cluster is stuck
waiting for the primary to be demoted to a designed primary, due to a
concurrent update on the Cluster CR by the operator and the instance
manager.

The update by the instance manager is now using the update method
instead of a patch, relying on the Kubernetes optimistic locking mechanism.

Closes: #5754

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 0f82afcb9eb36ba10519a30efc1d0286cb30e715)