chore(logging): avoid creating multiple loggers
With the latest version of controller-runtime the current logger can be
set in the base context and then inherited by all the controllers
belonging to the same manager.

This patch removes the duplicate loggers in the PostgreSQL instances and
in the PgBouncer Pods, reusing the appropriate one. This, in turn,
allows for a more consistent logging across the instance manager.

Closes: #5782

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>