fix(plugin): avoid jamming JSON logs with `logs` command (#5775)
`kubectl cnpg logs cluster` collects the log stream of all the Pods
belonging to the selected cluster and forwards their output to a final
destination stream with no coordination besides basic locking at the
byte level.

This resulted in multiple JSON objects being written at the same time,
which made the output garbled. Tools relying on the JSON-stream format
cannot parse that too.

With this patch, we coordinate the reading process, scan the stream line
by line, and write every separate line for the final stream.

Closes: #5769

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Ben Healey <bentastic27@gmail.com>