fix(failover): avoid failing over multiple times with unstable connectivity
This patch prevents the operator from failing over two times when a Pod
loses connectivity and doesn't notice the change of its current role
from primary to replica.

The first failover would happen when the operator notices the primary
Pod is not ready/present anymore. The second failover will happen if
the old primary comes back to life before the Kubelet timeout, with
the operator potentially failing over to the first one of the Pod list.

When this happens, we will wait for the Pod to understand its current
role.

Closes: #2513

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>