fix(sync): set synchronous_stanby_names also when no replica is reachable
Using the new synchronous replication configuration, if all the standby
pods in a cluster are unhealthy, the primary should not be able to
complete any commit.

However, if we find nothing to replicate on, the code dropped the
synchronous_standby_names setting.

This patch fills is out by considering all the Pods belonging to the
cluster as potential synchronous replica, indipendently of their status.

Closes: #5784

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>