chore(tests): Refactored backup and restore tests by separating backend-specific logic (#5735)
Previously, all backup and restore tests for different backends 
(MinIO, Azurite, Azure) were combined in a single file, making it
difficult to manage and maintain. This change separates the tests, 
creating individual files for each backend, with dedicated functions 
tailored to each backend's requirements. This improves readability
and organization of the code.

Closes #5632

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>