feat: make data durability configurable for sync replication (#5878)
This patch introduces an option to relax strict data durability
requirements within the new synchronous replication interface. By
setting `.spec.postgresql.synchronous.dataDurability` to `preferred`,
the operator limits the required number of synchronous standby replicas
to match the number of available instances, prioritizing self-healing
over strict data durability.

The `preferred` method requires both `standbyNamesPre` and
`standbyNamesPost` to be empty.

Closes: #5793

## Release Notes

Introduce the `dataDurability` option in the `.spec.postgresql.synchronous` stanza to control the trade-off between data safety and availability for synchronous replication. The option can be set to `required` or `preferred`, with the default being `required` if not specified.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>