fix: use optimistic locking when updating conditions
Kubernetes does not implement strategic merge for CRDs, and every JSON
merge patch will replace the condition set with a new one.

It is possible for a proposed patch to start from a Cluster which is not
up-to-date, and in that cases the conditions will be reverted back with
an older status.

This patch fixes this race condition by encapsulating this operator with
a merge patch that requires optimistic locking and retrying the
conditions update when needed.

Fixes: #6317

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>