fix: use optimistic locking when updating conditions (#6328)
Kubernetes does not implement strategic merge for CRDs, and every JSON
merge patch will replace the condition set with a new one.

It is possible for a proposed patch to start from a Cluster that is not
up-to-date, and in that case, the conditions will be reverted back to an
older status.

This patch fixes this race condition by encapsulating this operator with
a merge patch that requires optimistic locking and retrying the
conditions update when needed.

Fixes: #6317

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 67cc5473f3ca30f47d12860a01d57d29b2c47039)