fix: slow configuration update when decreasing max_connections
In case of a postgresql configuration update, in case of a decrease of a
hot standby sensible parameter with an unsupervised method, the instance
manager won't trigger a reconciliation loop of the operator after the
primary server has been restarted.

This patch is doing that by patching the status with a known phase and
reason.

Fix: #6409

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>