fix: consistent threshold calculation for probes (#6656)
In version 1.25.0, we introduced an inconsistent behavior in determining
the default value of the standard probe knobs when a stanza under
`.spec.probes` is defined (#6266).

This patch rectifies that behavior by allowing users to override any of
the settings, including `failureThreshold`. When `failureThreshold` is
not specified in the startup probe, its value is calculated by dividing
`.spec.startupDelay` by `periodSeconds` (which defaults to 10 and is now
overridable). The same principle applies to the liveness probe with the
`.spec.livenessProbeTimeout` option.

Closes: #6655

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>