chore(deps): pin genref version (#6871)
A recent commit in
https://github.com/kubernetes-sigs/reference-docs/genref introduced
changes to the generated documentation, causing the "Verify API doc is
up to date" checks to fail across all PRs.

This commit pins the `genref` dependency to the previous known working
commit. This will allow us to later integrate it with Renovate for
proper version management and controlled upgrades.

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>