fix: ensure operator version is reconciled (#6496)
In the current code, the `operatorVersion` annotation is set manually
when creating resources. This has led to several consistency issues, as
it is not always invoked.

This patch proposes a new approach that integrates the `operatorVersion`
reconciliation within the `GetFixedInheritedAnnotations` function. This
method offers several advantages:

- The function is already called throughout the code, making it easy to
implement the change consistently.
- It simplifies code maintenance since the `operatorVersion` no longer
requires a separate handling approach from the developer's perspective.

Closes #6457 

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>