fix: prevent leaking connection when snapshot backup fail (#6879)
The instance manager allocates a PostgreSQL connection every time a
volume snapshot backup need to be taken and then runs
`pg_backup_start` on it.  That connection will stay up until the backup
is done.

If the snapshotting process fails, the backup will be marked as failed
but the connection won't be closed.

This patch ensures that a successive backup is able to proceed correctly
by closing the previous connection.

A successive commit will allow the operator to close that stale commit
even when there's no successive backup.

Partially closes #6761 

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>