docs: add missing string in files names of the report command (#6866)
There was  a string missing, `(secret)` in the documentation when
referencing the files created by the report command in the plugin.

This was added to the output examples and the `head` command
that show part of the files.

Signed-off-by: Pierrick <pierrick.chovelon@dalibo.com>
(cherry picked from commit 627e8d31415a363cae0c0dea1b640a8f675ba0c1)