test: fix race on drain e2e (#6972)
If the storage was not local to the node, it was possible for the
drained pod to be recreated on a different node before we were able to
verify its pending state. We now consider a good behaviour to be
"pending" or already running on a different node.

Closes #6971

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
(cherry picked from commit 5bc22e7cee562b41ff9e354c37d25c8eab58f858)