fix: delay rollout until every Pods received the same configuration (#6991)
This patch delays the rollout of Pods to the point where every instance
manager received the latest PostgreSQL configuration and applied it.

This prevents the operator from switching over two times to apply the
same configuration when the primary received the configuration and,
immediately after that, a reconciliation loop has been triggered on the
cluster.

Fixes #6956

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>