chore(webserver,update): retrieve instance architecture using Go runtime (#6884)
The instance architecture is now determined using the Go runtime's
`runtime.GOARCH` variable, eliminating the need for an SQL connection.
This enhancement improves efficiency by reducing unnecessary database
interactions and simplifies the codebase by leveraging Go's built-in
capabilities.

This change is internal and does not affect the existing functionality
of the system.

Relates to #6599  #6761

Signed-off-by: Jeff Mealo <jeffm@gisual.com>
Signed-off-by: Jeff Mealo <jmealo@protonmail.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>