feat: startup and readiness probes for replicas (#6623)
Extend the startup and readiness probes configured through the
`.spec.probes.startup` and `.spec.probes.readiness` sections by adding
two additional parameters:

- `type`: Defines the criteria for considering the probe successful.
Accepted values include:
- `pg_isready`: This setting marks the probe as successful when the
`pg_isready` command exits with a status of `0`. This is the default for
both primary instances and replicas.
- `query`: This setting marks the probe as successful when a basic query
is executed locally on the `postgres` database.
- `streaming`: This setting marks the probe successful when the replica
starts streaming from its source and meets the specified lag
requirements (details below).
- `lag`: Specifies the maximum acceptable replication lag, measured in
bytes (expressed using Kubernetes quantities). This parameter is
applicable only when `type` is set to `streaming`. If the `lag`
parameter is not specified, the replica is considered successfully
started/ready as soon as it begins streaming.

Consequently, the liveness probe has been streamlined to verify solely
that the instance manager is operational, without monitoring the
underlying PostgreSQL instance.

Closes: #6621 

## Release Notes

**Improved Startup and Readiness Probes for Replicas**: Enhanced support
for Kubernetes startup and readiness probes in PostgreSQL instances,
providing greater control over replicas based on the streaming lag.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@gmail.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>