refactor: move the multi-container logging to a dedicated package (#6809)
This pull request refactors the code that iterates over containers when
logging instance pods for the `kubectl cnpg report` plugin command. The
refactoring involves moving the multi-container logging logic to a
dedicated package, which improves the codebase's organization and
maintainability. Additionally, new tests have been added to ensure the
functionality is thoroughly covered.

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>